+++ lib/fmcomms2_source_impl.h
@@ -41,9 +41,10 @@ namespace gr {
 
      public:
       fmcomms2_source_impl(const std::string &host,
-		    double frequency, double samplerate, double bandwidth,
+		    unsigned long long frequency,
+		    unsigned long samplerate, unsigned long bandwidth,
 		    bool ch1_en, bool ch2_en, bool ch3_en, bool ch4_en,
-		    unsigned int buffer_size, bool quadrature, bool rfdc,
+		    unsigned long buffer_size, bool quadrature, bool rfdc,
 		    bool bbdc, const char *gain1, double gain1_value,
 		    const char *gain2, double gain2_value,
 		    const char *rf_port_select);
